+++ jedi/tests/test_get_signature.py
@@ -147,6 +147,7 @@ CONSTRUCTOR_PARAMS = [
                 "right_motor: Motor",
                 "wheel_diameter: Number",
                 "axle_track: Number",
+                "positive_direction: Direction=Direction.CLOCKWISE",
             ]
         ],
     ),