+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Unreleased
 
 ### Added
+- Documented `positive_direction` in `DriveBase`.
+- Document deceleration setter in `DriveBase`.
 - Documented ``integral_deadzone`` in ``Control.pid()``.
 - Documented ``Motor.model``. This can be used to view the estimated motor
   state and change its settings.